Move user address tests out of test_conditions and into test_context module.
